M  Cargo.toml, M  src/constants.rs, M  src/evaluate.rs, M  src/lib.rs, M  src/polyglot.rs, M  src/search.rs, M  src/search_controller.rs, M  src/sort.rs, M  src/utils/extension_traits.rs, M  src/utils/lazy_static.rs

+++ src/constants.rs
@@ -12,7 +12,7 @@ pub mod types {
     pub type Ply = usize;
     pub type Depth = i8;
     pub type Score = i16;
-    pub type MoveWeight = i64;
+    pub type MoveWeight = i32;
     pub type NumMoves = u16;
     pub type Spin = u128;
     #[cfg(feature = "engine")]
